pyproject.toml
[project]
name = "tandarunner"
version = "24.4.13"
description = "A nice description"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "django==5.1b1",
  "django-allauth[socialaccount]",
  "altair",
  "pandas",
  "numpy",
  "whitenoise",
  "channels[daphne]",
  "django-debug-toolbar",
  "markdown",
  "Twisted[tls,http2]",
  "diskcache",
  "django-extensions",
  "pydantic-ai-slim[openrouter]>=1.70.0",
  "duckdb",
  "icalendar>=7.0.3",
]

[dependency-groups]
dev = [
  "pytest",
  "pytest-cov",
  "ruff",
  "ipykernel",
  "jupyter-black",
  "ty",
  "deptry",
  "djlint>=1.36.4",
]

[tool.ruff]
exclude = [".env", ".venv", "venv", "notebooks", "research"]
line-length = 79

[tool.ruff.lint]
ignore = ["E501"]
select = ["E", "F", "I", "W"]
fixable = ["I", "F"]

[tool.ty.src]
exclude = ["notebooks", "research"]

[tool.ty.rules]
unresolved-attribute = "ignore"
unresolved-import = "ignore"
invalid-method-override = "warn"
invalid-return-type = "warn"

[tool.deptry]
extend_exclude = ["notebooks", "research", "scripts"]

[tool.deptry.per_rule_ignores]
DEP002 = ["whitenoise", "daphne", "django-debug-toolbar", "django-extensions", "Twisted", "diskcache"]
DEP003 = ["asgiref", "requests", "pydantic", "icalendar"]

[tool.coverage.paths]
source = ["src"]

[tool.coverage.run]
branch = true
relative_files = true

[tool.coverage.report]
show_missing = true
fail_under = 80
